test: add tests for topBarComponent and bottomBarComponent to improve coverage
